+++ Source/AbstractPatch.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using HarmonyLib;
 using JetBrains.Annotations;
 using Replace_Stuff.NewThing;
+using RimWorld;
 using Verse;
 
 namespace Replace_Stuff_Compatibility
@@ -29,19 +30,34 @@ namespace Replace_Stuff_Compatibility
 
 		protected static ThingDef GetDatabaseThing(string name) => DefDatabase<ThingDef>.GetNamed(name);
 
-		private static Predicate<ThingDef> ListContainsThingDef(List<ThingDef> list)
+		protected static Predicate<ThingDef> ListContainsThingDef(List<ThingDef> list)
 		{
 			return product => list.Exists(n => n == product);
-		}
-
-		protected static void AddInterchangeableList(params ThingDef[] items)
-		{
-			NewThingReplacement.replacements.Add(new NewThingReplacement.Replacement(ListContainsThingDef(items.ToList())));
 		}
 		
 		protected static void AddInterchangeableList(List<ThingDef> items)
 		{
 			NewThingReplacement.replacements.Add(new NewThingReplacement.Replacement(ListContainsThingDef(items)));
 		}
+		
+		protected static void AddInterchangeableList(params ThingDef[] items) => AddInterchangeableList(items.ToList());
+		
+		protected static void AddInterchangeableWorkbenches(List<ThingDef> items)
+		{
+			Action<Thing, Thing> transferBills = (n, o) =>
+			{
+				Building_WorkTable newTable = n as Building_WorkTable;
+				Building_WorkTable oldTable = o as Building_WorkTable;
+
+				foreach (Bill bill in oldTable.BillStack)
+				{
+					newTable.BillStack.AddBill(bill);
+				}
+			};
+
+			NewThingReplacement.replacements.Add(new NewThingReplacement.Replacement(ListContainsThingDef(items), preAction: transferBills));
+		}
+		
+		protected static void AddInterchangeableWorkbenches(params ThingDef[] items) => AddInterchangeableWorkbenches(items.ToList());
 	}
 }