+++ Source/MultiModPatch.cs
@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using Replace_Stuff.NewThing;
+using RimWorld;
 using Verse;
 
 namespace Replace_Stuff_Compatibility
@@ -14,11 +16,35 @@ namespace Replace_Stuff_Compatibility
 
 		public static List<ThingDef> Lights = new List<ThingDef>() { };
 
+		public static List<ThingDef> Sunlamps = new List<ThingDef>() {GetDatabaseThing("SunLamp")};
+
+		public static List<ThingDef> EndTables = new List<ThingDef>() {GetDatabaseThing("EndTable")};
+
+		public static List<ThingDef> Dressers = new List<ThingDef>() {GetDatabaseThing("Dresser")};
+
+		public static List<ThingDef> TVs = new List<ThingDef>()
+		{
+			GetDatabaseThing("TubeTelevision"), GetDatabaseThing("FlatscreenTelevision"),
+			GetDatabaseThing("MegascreenTelevision")
+		};
+
 		protected override void AddItems()
 		{
-			AddInterchangeableList(Smelters);
-			AddInterchangeableList(Smithys);
+			// Allow all "plant growable items" to replace each other, and when they do attempt to set the growing plant type
+			NewThingReplacement.replacements.Add(new NewThingReplacement.Replacement(
+				building => typeof(IPlantToGrowSettable).IsAssignableFrom(building.thingClass),
+				postAction: (newItem, oldItem) =>
+				{
+					((IPlantToGrowSettable) newItem).SetPlantDefToGrow(((IPlantToGrowSettable) oldItem).GetPlantDefToGrow());
+				}));
+
+			AddInterchangeableWorkbenches(Smelters);
+			AddInterchangeableWorkbenches(Smithys);
 			AddInterchangeableList(Lights);
+			AddInterchangeableList(Sunlamps);
+			AddInterchangeableList(EndTables);
+			AddInterchangeableList(Dressers);
+			AddInterchangeableList(TVs);
 		}
 	}
 }