+++ Source/VanillaExpandedFurniture.cs
@@ -16,10 +16,6 @@ namespace Replace_Stuff_Compatibility
 
 		protected override void AddItems()
 		{
-			var tableLightEndTable = GetDatabaseThing("Table_LightEndTable");
-			var tableRoyalEndTable = GetDatabaseThing("Table_RoyalEndTable");
-			var tableRoyalDresser = GetDatabaseThing("Table_RoyalDresser");
-
 			var campfire = GetDatabaseThing("Campfire");
 			var stoneCampfire = GetDatabaseThing("Stone_Campfire");
 
@@ -29,9 +25,12 @@ namespace Replace_Stuff_Compatibility
 			MultiModPatch.Lights.Add(GetDatabaseThing("Light_ModernLamp"));
 			MultiModPatch.Lights.Add(GetDatabaseThing("Light_Streetlamp"));
 			MultiModPatch.Lights.Add(GetDatabaseThing("Table_LightEndTable"));
+			
+			MultiModPatch.EndTables.Add(GetDatabaseThing("Table_LightEndTable"));
+			MultiModPatch.EndTables.Add(GetDatabaseThing("Table_RoyalEndTable"));
+			
+			MultiModPatch.Dressers.Add(GetDatabaseThing("Table_RoyalDresser"));
 
-			AddInterchangeableList(BaseGameThingDefs.EndTable, tableLightEndTable, tableRoyalEndTable);
-			AddInterchangeableList(BaseGameThingDefs.Dresser, tableRoyalDresser);
 			AddInterchangeableList(campfire, stoneCampfire);
 			AddInterchangeableList(oldRadio, radio);
 		}